Move location of test data
