+++ sdk/tests/wasm.test.ts
@@ -1,5 +1,5 @@
 import { Address, PrivateKey, ViewKey, PrivateKeyCiphertext, Signature, RecordCiphertext, RecordPlaintext } from "@aleohq/wasm";
-import { seed, message, privateKeyString, viewKeyString, addressString, recordCiphertextString, foreignViewKeyString, recordPlaintextString } from './account-data';
+import { seed, message, privateKeyString, viewKeyString, addressString, recordCiphertextString, foreignViewKeyString, recordPlaintextString } from './data/account-data';
 
 describe('WASM Objects', () => {
     describe('Address', () => {
