+++ sdk/tsconfig.json
@@ -1,6 +1,6 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-  "include": ["src"],
+  "include": ["src", "src/models"],
   "exclude": ["node_modules", "tests"],
   "compilerOptions": {
     "module": "esnext",
